+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2023-12-08
+* add ability to force delete Application/Parent:
+  * r8s application delete --force
+  * r8s parent delete --force
+
 ## [3.0.1] - 2023-12-08
 * adapt shape rules API to new Application/Parent model